+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## **unreleased** - unreleased
 - Made Lisp code tags more useful: the HTTP method for a request spec can now be determined by the result of a code tag.
 - Renamed customizable variable `verb-show-headers-buffer` to `verb-auto-show-headers-buffer`.
+- Added `verb-last` variable.
+- Added `verb-headers-to-string` function.
 
 ## **1.0.0** - 2020-01-13
 - Initial release.